servers/compile_diagnostic_mcp/src/requirements.txt
fastmcp
regex